ENH: move flexible_vstack to __init__
